bookedForTimes key added to fakedata json, unnecessary animation removed from homepage, footer edited, service section edited, cards completed, apartment component moved outside of home
