+++ src/components/Home/Cards/SingleCard/SingleCard.css
@@ -1,0 +1,4 @@
+.apartment-card:hover {
+    box-shadow: 0 8px 12px 0 rgba(0,0,0,0.3);
+    cursor:pointer;
+}
