+++ src/components/Home/Cards/SingleCard/SingleCard.js
@@ -1,11 +1,12 @@
 import React from 'react';
 import Card from 'react-bootstrap/Card';
-import Button from 'react-bootstrap/Button';
 import Reveal from 'react-reveal/Reveal';
 import { useHistory } from 'react-router-dom';
+import './SingleCard.css';
+
 const SingleCard = (props) => {
     console.log(props);
-    const { apartmentName, id, image, apartmentPrice, type } = props.data;
+    const { apartmentName, id, size, monthlyRent, address, image, apartmentPrice, type } = props.data;
 
     const history = useHistory()
     const handleSelectApartment = id => {
@@ -14,15 +15,13 @@ const SingleCard = (props) => {
 
     return (
         <Reveal duration={3000}>
-            <Card onClick={() => handleSelectApartment(id)} style={{ width: '18rem', marginBlock: '1rem' }}>
-                <Card.Img variant="top" src={props.data.image} />
+            <Card onClick={() => handleSelectApartment(id)} className='apartment-card' style={{ width: '18rem', marginBlock: '1rem' }}>
+                <Card.Img variant="top" src={image} />
+                <Card.Title style={{textAlign: 'center', marginTop: '10px'}}>{apartmentName}</Card.Title>
                 <Card.Body>
-                    <Card.Title>Card Title</Card.Title>
-                    <Card.Text>
-                        Some quick example text to build on the card title and make up the bulk of
-                        the card's content.
-                </Card.Text>
-                    <Button variant="warning">Go somewhere</Button>
+                    <Card.Text>{type} | {size}</Card.Text>
+                    <Card.Text>{address}</Card.Text>
+                    <Card.Text>Apartment Price: {apartmentPrice} | Monthly Rent: {monthlyRent}</Card.Text>
                 </Card.Body>
             </Card>
         </Reveal>
