+++ src/components/Home/FooterCol/FooterCol.js
@@ -1,5 +1,4 @@
 import React from 'react';
-// import { Link } from 'react-router-dom';
 
 const FooterCol = (props) => {
     return (
